requirements.txt
asgiref==3.3.1
Django==2.1.15
django-mssql-backend==2.8.1
django-pyodbc-azure==2.1.0.0
djangorestframework==3.12.2
pyodbc==4.0.30
pytz==2020.5
sqlparse==0.4.1
